+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/ForOneLab/cert
 
-go 1.18
+go 1.21
 
 require (
 	k8s.io/api v0.25.4
